+++ demo/lib/queue.ex
@@ -21,8 +21,8 @@ defmodule Demo.Queue do
   end
 
   @impl Parent.GenServer
-  def handle_child_terminated(_id, _meta, _pid, reason, state) do
-    if reason == :normal do
+  def handle_child_terminated(info, state) do
+    if info.reason == :normal do
       IO.puts("job succeeded")
     else
       IO.puts("job failed")